M5 Phase 1: data model split and save/load
- PlanetState gets LocalWood/LocalWater; parkActive saves them,
  loadPlanet restores them so Economy.Wood/Water become per-planet
  local stockpiles on planet switch
- SystemEconomy struct added to World with WoodRate/WaterRate,
  WoodAllocPotential/WaterAllocPotential (init 1.0), WoodResearch/
  WaterResearch; survives save/load
- Economy.Potential changed from map[PotentialKind]int to
  map[PotentialKind]float64; spendable count is math.Floor(value);
  canAwaken, awakenPlanet, awardCompletionPotential, display sites
  all updated
- ProjectedRate/ProjectedWaterRate added to SystemPlanet
- SaveVersion bumped to 19→20
- Round-trip tests for all new fields

+++ internal/game/simtrace_test.go
@@ -433,10 +433,10 @@ func TestSimTraceTightGroveCompletion(t *testing.T) {
 		t.Errorf("Tight Grove: expected Completed=true after trace")
 	}
 	if got := w.Economy.Potential[PotentialForest]; got != 1 {
-		t.Errorf("Forest Potential after Tight Grove: got %d, want 1", got)
+		t.Errorf("Forest Potential after Tight Grove: got %.0f, want 1", got)
 	}
 	if got := w.Economy.Potential[PotentialWater]; got != 0 {
-		t.Errorf("Water Potential after Tight Grove: got %d, want 0 (no water field)", got)
+		t.Errorf("Water Potential after Tight Grove: got %.0f, want 0 (no water field)", got)
 	}
 	t.Logf("Tight Grove completed — AbstractRate: %.4f wood/sec", w.System.Planets[echoIdx].AbstractRate)
 }
@@ -457,7 +457,7 @@ func TestSimTraceLakewoodCompletion(t *testing.T) {
 		t.Errorf("Lakewood: expected Completed=true after trace")
 	}
 	if got := w.Economy.Potential[PotentialWater]; got != 1 {
-		t.Errorf("Water Potential after Lakewood: got %d, want 1", got)
+		t.Errorf("Water Potential after Lakewood: got %.0f, want 1", got)
 	}
 	// Verify lake-aware routing enabled workers to reach the island forest.
 	var islandField *ResourceField
@@ -672,7 +672,7 @@ func (r *waterPlanetCompletionRunner) Setup(w *World) {
 
 	// Grant + spend Potential to awaken the frontier (index 3).
 	for kind, cost := range planetAwakenCost(w, 3) {
-		w.Economy.Potential[kind] += cost
+		w.Economy.Potential[kind] += float64(cost)
 	}
 	awakenPlanet(w, 3)
 	switchToPlanet(w, 3)
